Iterations limit for test runs (#59)
`--iterations` parameter added to have the ability to set an iteration
limit for test runs as a complementary option to `--runtime`. If set,
the test will run only a given number of iterations. I


`--runtime` parameter is always taken into account as a safeguard to
prevent tests from running indefinitely. If both `--runtime` and
`--iterations` limits are defined, the test will stop when either limit
is reached first.